feat: Complete functionality to update a course from the edit view
To simplify the course mentor tracking, I removed the mentor as a DB
entity, and just made it an embedded object within the course entity.
Therefore, the mentor fields are part of the course table in the DB, but
are retrieved from the course object as a mentor object using a getter.

Related issue: #2