feat: Add ability to create a new course from course list or term view
I had to create a separate asynctask within the termRepository class
just for this feature.  For some reason, I was not able to get the
list of terms in the course edit activity with the LiveData.

I also removed the delete options from the edit activities for terms,
courses, and assessments.

Resolves: #21