feat: Allow for up to three mentors per course
This is a feature requirement that was not clearly stated, so it was
not present in the first submission.  To the resubmission quicker, I
just added the extra fields to the mentor class, rather than properly
change the DB structure to allow for multiple mentors in a normalized
way.  This could definitely be improved upon.

Related issue: #47